make expression-condition controller more generic
