+++ app/javascript/packs/application.js
@@ -17,7 +17,7 @@ import ConditionalController from '../src/controllers/conditional-controller.js'
 import ConditionalsController from '../src/controllers/conditionals-controller.js'
 import ExpressionController from '../src/controllers/expression-controller.js'
 import ExpressionsController from '../src/controllers/expressions-controller.js'
-import ExpressionConditionController from '../src/controllers/expression-condition-controller.js'
+import SelectionRevealController from '../src/controllers/selection-reveal-controller.js'
 
 Turbo.session.drive = false
 window.Stimulus = Application.start()
@@ -26,7 +26,7 @@ Stimulus.register("conditional", ConditionalController)
 Stimulus.register("conditionals", ConditionalsController)
 Stimulus.register("expression", ExpressionController)
 Stimulus.register("expressions", ExpressionsController)
-Stimulus.register("expression-condition", ExpressionConditionController)
+Stimulus.register("selection-reveal", SelectionRevealController)
 
 const { EditableContent } = require('../src/web-components/editable-content');
 const { ElasticTextarea } = require('../src/web-components/elastic-textarea');
