+++ app/views/api/conditional_contents/_expression_condition.html.erb
@@ -1,6 +1,6 @@
 <turbo-frame id="conditional_<%= conditional_index %>_expression_<%= expression_index %>" data-expression-target="condition">
   <% if expression.component.present? && expression.component_supported? %>
-    <div class="expression__condition" data-controller="expression-condition" data-expression-condition-matches-value="<%= expression.answered_operator_values.to_json %>">
+    <div class="expression__condition" data-controller="selection-reveal" data-selection-reveal-matches-value="<%= expression.answered_operator_values.to_json %>">
         <%= f.select :operators, expression.operators,
           {
             include_blank: false,
@@ -9,8 +9,8 @@
           {
             class: "govuk-select expression__operator #{ expression.errors[:operator].any? ? 'govuk-select--error' : '' }",
             data: {
-              expression_condition_target: 'operator',
-              action: "expression-condition#toggleAnswer"
+              selection_reveal_target: 'input',
+              action: "selection-reveal#toggle"
             },
             name: expression.name_attr(
               conditional_index: conditional_index,
@@ -33,7 +33,9 @@
         },
         {
           class: "govuk-select expression__answer #{ expression.errors[:answer].any? ? 'govuk-select--error' : '' }",
-          'data-expression-condition-target': 'answer',
+          data: {
+            selection_reveal_target: 'reveal',
+          },
           name: expression.name_attr(
             conditional_index: conditional_index,
             expression_index: expression_index,
